Added functionality for building RAID0 Block Volume cache on Workers for moving data to/from Object Storage (Distcp).
This is off by default, and built into the ORM Schema template as a boolean to enable.

This feature uses conditionals with the Worker module to build extra block volumes, and change device mappings for data volumes when enabled.   Also built a detection flag into the boot script to toggle RAID array setup based on whether the boolean is true/false.