Updated Cloudera versions to refelect what is currently available in the public archive.
Note that this repository is not maintained on a frequent basis - these values may need to be updated quarterly to reflect what is not behind the Cloudera paywall.  Refer to https://archive.cloudera.com/cdh7/ and ensure both schema.yaml and variables.tf reflect the appropriate version which is currently available.

Failure to download parcels during deployment is a symptom of this version URL being incorrect - it's dynamically generated based on the values in the two files mentioned above.

Also added the Deploy to OCI magic button, and updated OS images to refer to latest.